pyproject.toml
[project]
name = "scholar-vision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.134.0",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings>=2.13.1",
    "requests>=2.32.5",
    "aiofiles>=24.1.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.5.0",
    "shap>=0.46.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
]
